# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Smart City Digital Citizen Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List complaints
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, resolved, closed]
        - name: category
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated complaints list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintListResponse"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Submit a new complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Complaint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get complaint by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Complaint details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintUpdate"
      responses:
        "200":
          description: Updated complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{id}/ai-analysis:
    post:
      operationId: analyzeComplaint
      tags: [complaints, ai]
      summary: Run AI analysis on a complaint
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIAnalysis"
...
  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List certificate applications
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [birth, death]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, under_review, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Certificate list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateListResponse"
    post:
      operationId: applyCertificate
      tags: [certificates]
      summary: Apply for a certificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateInput"
      responses:
        "201":
          description: Certificate application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
  /certificates/{id}:
    get:
      operationId: getCertificate
      tags: [certificates]
      summary: Get certificate application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Certificate details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
    patch:
      operationId: updateCertificate
      tags: [certificates]
      summary: Update certificate status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateUpdate"
      responses:
        "200":
          description: Updated certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"

  # ─────────────────────────────── GARBAGE ───────────────────────────────
...
  /ai/chat:
    post:
      operationId: sendChatMessage
      tags: [ai]
      summary: Send a message to the AI citizen assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
  /ai/chat/history:
    get:
      operationId: getChatHistory
      tags: [ai]
      summary: Get AI chat history for current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [citizen, admin, super_admin]
        createdAt:
          type: string
          format: date-time
      required: [id, clerkId, email, firstName, lastName, role, createdAt]
...
    Complaint:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [electricity, water, roads, garbage, transport, street_lights, drainage, public_safety, environment, other]
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [pending, in_progress, resolved, closed]
        location:
          type: string
        address:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        userId:
          type: string
        departmentId:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        aiCategory:
          type: ["string", "null"]
        aiPriority:
          type: ["string", "null"]
        aiConfidence:
          type: ["number", "null"]
        estimatedResolution:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [id, title, description, category, priority, status, location, userId, createdAt, updatedAt, user]
    ComplaintInput:
      type: object
      properties:
        title:
          type: string
          minLength: 5
        description:
          type: string
          minLength: 20
        category:
          type: string
          enum: [electricity, water, roads, garbage, transport, street_lights, drainage, public_safety, environment, other]
        location:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        departmentId:
          type: string
      required: [title, description, category, location]
    ComplaintUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        assignedTo:
          type: string
        departmentId:
          type: string
        estimatedResolution:
          type: string
    ComplaintListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Complaint"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]
...
    AIAnalysis:
      type: object
      properties:
        complaintId:
          type: string
        category:
          type: string
        confidence:
          type: number
        priority:
          type: string
        reasoning:
          type: string
        estimatedResolutionTime:
          type: string
        suggestedDepartment:
          type: string
      required: [complaintId, category, confidence, priority, reasoning, estimatedResolutionTime, suggestedDepartment]

    # ─── Property Tax ───
...
    Certificate:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [birth, death]
        status:
          type: string
          enum: [pending, under_review, approved, rejected]
        applicantName:
          type: string
        applicantRelation:
          type: string
        subjectName:
          type: string
        subjectDateOfBirth:
          type: ["string", "null"]
          format: date
        subjectDateOfDeath:
          type: ["string", "null"]
          format: date
        placeOfEvent:
          type: string
        remarks:
          type: ["string", "null"]
        adminRemarks:
          type: ["string", "null"]
        certificateNumber:
          type: ["string", "null"]
        approvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, type, status, applicantName, applicantRelation, subjectName, placeOfEvent, createdAt, updatedAt]
    CertificateInput:
      type: object
      properties:
        type:
          type: string
          enum: [birth, death]
        applicantName:
          type: string
        applicantRelation:
          type: string
        subjectName:
          type: string
        subjectDateOfBirth:
          type: string
          format: date
        subjectDateOfDeath:
          type: string
          format: date
        placeOfEvent:
          type: string
        remarks:
          type: string
      required: [type, applicantName, applicantRelation, subjectName, placeOfEvent]
    CertificateUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, under_review, approved, rejected]
        adminRemarks:
          type: string
        certificateNumber:
          type: string
    CertificateListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Certificate"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]

    # ─── Garbage ───
...
    ChatMessageInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
        sessionId:
          type: ["string", "null"]
      required: [message]
    ChatResponse:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
        sessionId:
          type: string
        timestamp:
          type: string
          format: date-time
        suggestions:
          type: array
          items:
            type: string
      required: [id, message, sessionId, timestamp, suggestions]
    ChatMessage:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        sessionId:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [id, role, content, sessionId, timestamp]

    # ─── Departments ───
...