# lib/api-zod/src/generated/types/complaintListResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Smart City Digital Citizen Portal API
 * OpenAPI spec version: 1.0.0
 */
import type { Complaint } from './complaint';
import type { PaginationMeta } from './paginationMeta';

export interface ComplaintListResponse {
  data: Complaint[];
  pagination: PaginationMeta;
}